requirements.txt
annotated-types==0.7.0
anyio==4.10.0
click==8.2.1
dotenv==0.9.9
exceptiongroup==1.3.0
faiss-cpu==1.12.0
fastapi==0.116.1
h11==0.16.0
idna==3.10
ImageHash==4.3.2
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
pydantic==2.11.7
pydantic_core==2.33.2
pytesseract==0.3.13
python-dotenv==1.1.1
python-multipart==0.0.20
PyWavelets==1.8.0
scipy==1.15.3
sniffio==1.3.1
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.35.0
